+++ crates/zng-app/src/widget/base.rs
@@ -932,6 +932,13 @@ pub mod node {
                 visitor(0, &mut self.child)
             }
 
+            fn try_for_each_child(
+                &mut self,
+                visitor: &mut dyn FnMut(usize, &mut UiNode) -> std::ops::ControlFlow<zng_var::BoxAnyVarValue>,
+            ) -> std::ops::ControlFlow<zng_var::BoxAnyVarValue> {
+                visitor(0, &mut self.child)
+            }
+
             fn par_each_child(&mut self, visitor: &(dyn Fn(usize, &mut UiNode) + Sync)) {
                 visitor(0, &mut self.child)
             }